Opening Twitter URLs in SafariVC
Was there any reason to use an inline UIWebView (not even WKWebView 😁)?
The only flexibility that offers (afaik) is to display a custom title.

As this wasn't that useful when showing a Twitter page - and being
logged in to Twitter of Safari would be useful - I've changed those to
use SafariVC